pyproject.toml
[tool.poetry]
name = "nomiatest"
version = "0.1.0"
description = ""
authors = ["Vladislav-Shi <vlad.shirobockow@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4"
uuid = "^1.30"
django-nested-admin = "^4.0.2"
django-bootstrap-v5 = "^1.0.11"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
